requirements.txt
google-api-python-client
requests
python-dotenv
tensorflow
opencv-python
numpy
Pillow
Flask==2.2.5